pyproject.toml
[project]
name = "fraud-detection-platform"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "psycopg2-binary",
    "pydantic[email]",
    "pydantic-settings",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "python-dotenv"
]

[tool.vercel.scripts]
app = "api.main:app"
